+++ ObsStrawket/DataTypes/Frames/5_Event.cs
@@ -6,7 +6,7 @@ namespace ObsStrawket.DataTypes {
   /// <summary>
   /// Sent from: obs-websocket<br />
   /// Sent to: All subscribed and identified clients<br />
-  /// Description: An event coming from OBS has occured. E.g. scene switched, source muted.
+  /// Description: An event coming from OBS has occurred. E.g. scene switched, source muted.
   /// </summary>
   public interface IObsEvent : IOpCodeMessage {
     /// <summary>
@@ -23,7 +23,7 @@ namespace ObsStrawket.DataTypes {
   /// <summary>
   /// Sent from: obs-websocket<br />
   /// Sent to: All subscribed and identified clients<br />
-  /// Description: An event coming from OBS has occured. E.g. scene switched, source muted.
+  /// Description: An event coming from OBS has occurred. E.g. scene switched, source muted.
   /// </summary>
   [MessagePackObject]
   public class ObsEvent : IObsEvent {
@@ -49,8 +49,8 @@ namespace ObsStrawket.DataTypes {
   /// <summary>
   /// Sent from: obs-websocket<br />
   /// Sent to: All subscribed and identified clients<br />
-  /// Description: An event coming from OBS has occured. E.g. scene switched, source muted.<br />
-  /// Unidentifiable event (E.g. custom broadcast event) will be this.
+  /// Description: An event coming from OBS has occurred. E.g. scene switched, source muted.<br />
+  /// Unidentifiable event will be this.
   /// </summary>
   [MessagePackObject]
   public sealed class RawEvent : IObsEvent {